pyproject.toml
[project]
name = "mcp-interviewer"
version = "0.0.1"
description = "A tool for analyzing and testing MCP (Model Context Protocol) servers and their tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autogen-core",
    "autogen-ext[mcp]",
    "loguru>=0.7.3",
    "mcp>=1.10.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
mcp-interviewer = "mcp_interviewer.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

trapi = ["magentic-common[trapi-autogen]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_interviewer"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
testpaths = ["tests"]

[tool.uv.sources]
magentic-common = { git = "ssh://git@github.com/microsoft/magentic-common.git" }
autogen-core = { git = "https://github.com/tylerpayne/autogen.git", subdirectory = "python/packages/autogen-core", rev = "mcp-workbench-client" }
autogen-ext = { git = "https://github.com/tylerpayne/autogen.git", subdirectory = "python/packages/autogen-ext", rev = "mcp-workbench-client" }

